+++ inject/src/test/kotlin/fr/xgouchet/elmyr/inject/KotlinForgeryInjectorTest.kt
@@ -10,9 +10,15 @@ import fr.xgouchet.elmyr.inject.dummy.KotlinInjected
 import fr.xgouchet.elmyr.inject.dummy.KotlinInjectedChild
 import fr.xgouchet.elmyr.inject.dummy.KotlinInjectedGenerics
 import fr.xgouchet.elmyr.inject.dummy.KotlinInjectedImmutableVal
+import fr.xgouchet.elmyr.inject.dummy.KotlinInjectedInvalidPrimitives
 import fr.xgouchet.elmyr.inject.dummy.KotlinInjectedMissingFactory
+import fr.xgouchet.elmyr.inject.dummy.KotlinInjectedPrimitives
+import fr.xgouchet.elmyr.inject.dummy.KotlinInjectedRegex
+import fr.xgouchet.elmyr.inject.dummy.KotlinInjectedStrings
+import fr.xgouchet.elmyr.inject.dummy.KotlinInjectedUnknownAnnotation
 import fr.xgouchet.elmyr.inject.dummy.KotlinInjectedUnknownGeneric
 import org.assertj.core.api.Assertions.assertThat
+import org.assertj.core.data.Offset
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
@@ -112,6 +118,15 @@ class KotlinForgeryInjectorTest {
     }
 
     @Test
+    fun failsWhenInvalidPrimitiveParameters() {
+        val injected = KotlinInjectedInvalidPrimitives()
+
+        assertThrows<ForgeryException> {
+            injector.inject(forge, injected)
+        }
+    }
+
+    @Test
     fun failsWhenMissingFactory() {
         val injected = KotlinInjectedMissingFactory()
 
@@ -136,5 +151,97 @@ class KotlinForgeryInjectorTest {
         assertThrows<ForgeryInjectorException> {
             injector.inject(forge, injected)
         }
+    }
+
+    @Test
+    fun ignoresWhenUnknownAnnotation() {
+        val injected = KotlinInjectedUnknownAnnotation()
+
+        injector.inject(forge, injected)
+
+        assertThat(injected.retrieveFooOrNull()).isNull()
+    }
+
+    @Test
+    fun injectPrimitiveInt() {
+        val injected = KotlinInjectedPrimitives()
+
+        injector.inject(forge, injected)
+
+        assertThat(injected.publicInt).isNotEqualTo(0)
+        assertThat(injected.publicRangeInt).isBetween(3, 42)
+        assertThat(injected.publicGaussianInt).isCloseTo(1337, Offset.offset(20))
+        assertThat(injected.internalInt).isNotEqualTo(0)
+        assertThat(injected.retrieveProtectedInt()).isNotEqualTo(0)
+        assertThat(injected.retrievePrivateInt()).isNotEqualTo(0)
+    }
+
+    @Test
+    fun injectPrimitiveLong() {
+        val injected = KotlinInjectedPrimitives()
+
+        injector.inject(forge, injected)
+
+        assertThat(injected.publicLong).isNotEqualTo(0)
+        assertThat(injected.publicRangeLong).isBetween(3, 42)
+        assertThat(injected.publicGaussianLong).isCloseTo(1337L, Offset.offset(20L))
+        assertThat(injected.internalLong).isNotEqualTo(0)
+        assertThat(injected.retrieveProtectedLong()).isNotEqualTo(0)
+        assertThat(injected.retrievePrivateLong()).isNotEqualTo(0)
+    }
+
+    @Test
+    fun injectPrimitiveFloat() {
+        val injected = KotlinInjectedPrimitives()
+
+        injector.inject(forge, injected)
+
+        assertThat(injected.publicFloat).isNotEqualTo(0)
+        assertThat(injected.publicRangeFloat).isBetween(3f, 42f)
+        assertThat(injected.publicGaussianFloat).isCloseTo(1337f, Offset.offset(20f))
+        assertThat(injected.internalFloat).isNotEqualTo(0)
+        assertThat(injected.retrieveProtectedFloat()).isNotEqualTo(0)
+        assertThat(injected.retrievePrivateFloat()).isNotEqualTo(0)
+    }
+
+    @Test
+    fun injectPrimitiveDouble() {
+        val injected = KotlinInjectedPrimitives()
+
+        injector.inject(forge, injected)
+
+        assertThat(injected.publicDouble).isNotEqualTo(0)
+        assertThat(injected.publicRangeDouble).isBetween(3.0, 42.0)
+        assertThat(injected.publicGaussianDouble).isCloseTo(1337.0, Offset.offset(20.0))
+        assertThat(injected.internalDouble).isNotEqualTo(0)
+        assertThat(injected.retrieveProtectedDouble()).isNotEqualTo(0)
+        assertThat(injected.retrievePrivateDouble()).isNotEqualTo(0)
+    }
+
+    @Test
+    fun injectStrings() {
+        val injected = KotlinInjectedStrings()
+
+        injector.inject(forge, injected)
+
+        assertThat(injected.publicAplhaString).matches("[a-zA-Z]+")
+        assertThat(injected.publicAplhaLowerString).matches("[a-z]+")
+        assertThat(injected.publicAplhaUpperString).matches("[A-Z]+")
+        assertThat(injected.publicAplhaNumString).matches("[a-zA-Z0-9]+")
+        assertThat(injected.publicDigitsString).matches("[0-9]+")
+        assertThat(injected.retrieveProtectedHexString()).matches("[a-fA-F0-9]+")
+        assertThat(injected.retrievePrivateWhitespaceString()).matches("\\s+")
+    }
+
+    @Test
+    fun injectRegex() {
+        val injected = KotlinInjectedRegex()
+
+        injector.inject(forge, injected)
+
+        assertThat(injected.publicAlphaString).matches("[a-z]+")
+        assertThat(injected.internalDigitsString).matches("[0-9]+")
+        assertThat(injected.retrieveProtectedBase64String()).matches("([a-zA-z0-9]{4})*[a-zA-z0-9]{2}==")
+        assertThat(injected.retrievePrivatePhoneNumber()).matches("0\\d(-\\d\\d){4}")
     }
 }