+++ inject/src/test/kotlin/fr/xgouchet/elmyr/inject/dummy/KotlinInjectedGenerics.kt
@@ -8,7 +8,4 @@ open class KotlinInjectedGenerics {
     @Forgery lateinit var publicFooSet: Set<Foo>
     @Forgery lateinit var publicFooMap: Map<Foo, Bar>
     @Forgery lateinit var publicFooColletion: Collection<Foo>
-
-    private fun doSomething() {
-    }
 }