+++ inject/src/test/kotlin/fr/xgouchet/elmyr/inject/dummy/KotlinInjectedStrings.kt
@@ -1,0 +1,45 @@
+package fr.xgouchet.elmyr.inject.dummy
+
+import fr.xgouchet.elmyr.Case
+import fr.xgouchet.elmyr.annotation.StringForgery
+import fr.xgouchet.elmyr.annotation.StringForgeryType
+
+open class KotlinInjectedStrings {
+
+    @StringForgery(StringForgeryType.ALPHABETICAL)
+    lateinit var publicAplhaString: String
+
+    @StringForgery(StringForgeryType.ALPHABETICAL, Case.LOWER)
+    lateinit var publicAplhaLowerString: String
+
+    @StringForgery(StringForgeryType.ALPHABETICAL, Case.UPPER)
+    lateinit var publicAplhaUpperString: String
+
+    @StringForgery(StringForgeryType.ALPHA_NUMERICAL)
+    lateinit var publicAplhaNumString: String
+
+    @StringForgery(StringForgeryType.NUMERICAL)
+    lateinit var publicDigitsString: String
+
+    @StringForgery(StringForgeryType.ASCII)
+    internal lateinit var internalAsciiString: String
+
+    @StringForgery(StringForgeryType.ASCII_EXTENDED)
+    internal lateinit var internalAsciiExtString: String
+
+    @StringForgery(StringForgeryType.HEXADECIMAL)
+    protected lateinit var protectedHexString: String
+
+    @StringForgery(StringForgeryType.WHITESPACE)
+    private lateinit var privateWhitespaceString: String
+
+    fun retrieveProtectedHexString(): String = protectedHexString
+
+    fun retrievePrivateWhitespaceString(): String {
+        doSomething()
+        return privateWhitespaceString
+    }
+
+    private fun doSomething() {
+    }
+}